requirements.txt
crewai
langchain
openai
anthropic
google-generativeai
deepseek
pyyaml
docker
requests
fastapi
uvicorn
playwright
jinja2
pandas
pydantic
python-wappalyzer
lxml
beautifulsoup4